+++ pajbot/managers/schedule.py
@@ -1,57 +1,66 @@
+from typing import Any, Callable, Iterable, Mapping, Optional
+
 import datetime
 import logging
 
 from pajbot import utils
 
+from apscheduler.job import Job
 from apscheduler.schedulers.background import BackgroundScheduler
+from apscheduler.schedulers.base import BaseScheduler
 
 log = logging.getLogger(__name__)
 
 
 class ScheduledJob:
-    def __init__(self, job):
+    def __init__(self, job: Job) -> None:
         self.job = job
 
-    def pause(self, *args, **kwargs):
-        if self.job:
-            self.job.pause(*args, **kwargs)
+    def pause(self) -> None:
+        self.job.pause()
 
-    def resume(self, *args, **kwargs):
-        if self.job:
-            self.job.resume(*args, **kwargs)
+    def resume(self) -> None:
+        self.job.resume()
 
-    def remove(self, *args, **kwargs):
-        if self.job:
-            self.job.remove(*args, **kwargs)
+    def remove(self) -> None:
+        self.job.remove()
 
 
 class ScheduleManager:
-    base_scheduler = None
+    base_scheduler = BackgroundScheduler(daemon=True)
 
     @staticmethod
-    def init():
-        if not ScheduleManager.base_scheduler:
-            ScheduleManager.base_scheduler = BackgroundScheduler(daemon=True)
-            ScheduleManager.base_scheduler.start()
+    def init() -> None:
+        if ScheduleManager.base_scheduler.running:
+            log.warning("ScheduleManager had its init function called twice!!!!")
+            return
+
+        print(type(ScheduleManager.base_scheduler))
+        ScheduleManager.base_scheduler.start()
 
     @staticmethod
-    def execute_now(method, args=[], kwargs={}, scheduler=None):
+    def execute_now(
+        method: Callable[..., Any],
+        args: Optional[Iterable[Any]] = None,
+        kwargs: Optional[Mapping[str, Any]] = None,
+        scheduler: Optional[BaseScheduler] = None,
+    ) -> ScheduledJob:
         if scheduler is None:
             scheduler = ScheduleManager.base_scheduler
-
-        if scheduler is None:
-            raise ValueError("No scheduler available")
 
         job = scheduler.add_job(method, "date", run_date=utils.now(), args=args, kwargs=kwargs)
         return ScheduledJob(job)
 
     @staticmethod
-    def execute_delayed(delay, method, args=[], kwargs={}, scheduler=None):
+    def execute_delayed(
+        delay: float,
+        method: Callable[..., Any],
+        args: Optional[Iterable[Any]] = None,
+        kwargs: Optional[Mapping[str, Any]] = None,
+        scheduler: Optional[BaseScheduler] = None,
+    ) -> ScheduledJob:
         if scheduler is None:
             scheduler = ScheduleManager.base_scheduler
-
-        if scheduler is None:
-            raise ValueError("No scheduler available")
 
         job = scheduler.add_job(
             method, "date", run_date=utils.now() + datetime.timedelta(seconds=delay), args=args, kwargs=kwargs
@@ -59,12 +68,16 @@ class ScheduleManager:
         return ScheduledJob(job)
 
     @staticmethod
-    def execute_every(interval, method, args=[], kwargs={}, scheduler=None, jitter=None):
+    def execute_every(
+        interval: float,
+        method: Callable[..., Any],
+        args: Optional[Iterable[Any]] = None,
+        kwargs: Optional[Mapping[str, Any]] = None,
+        scheduler: Optional[BaseScheduler] = None,
+        jitter: Optional[float] = None,
+    ) -> ScheduledJob:
         if scheduler is None:
             scheduler = ScheduleManager.base_scheduler
-
-        if scheduler is None:
-            raise ValueError("No scheduler available")
 
         job = scheduler.add_job(method, "interval", seconds=interval, args=args, kwargs=kwargs, jitter=jitter)
         return ScheduledJob(job)